# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminUpdateUserRequest:
    properties:
      group_id:
        type: string
      is_blocked:
        type: boolean
      role:
        type: string
    type: object
  dto.AdminUpsertGroupRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
      success:
        type: boolean
    type: object
  dto.GroupResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.PaginationResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  dto.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      group:
        $ref: '#/definitions/dto.GroupResponse'
      id:
        type: string
      is_blocked:
        type: boolean
      last_name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: HTTP API for Stud Hub backend.
  title: Stud Hub API
  version: "1.0"
paths:
  /admin/groups:
    get:
      description: Returns all groups.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GroupResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List groups (admin)
      tags:
      - Admin Groups
    post:
      consumes:
      - application/json
      description: Creates a new group.
      parameters:
      - description: Create group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpsertGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Create group (admin)
      tags:
      - Admin Groups
  /admin/groups/{id}:
    delete:
      description: Deletes group by ID.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Delete group (admin)
      tags:
      - Admin Groups
    get:
      description: Returns group details by ID.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get group by ID (admin)
      tags:
      - Admin Groups
    patch:
      consumes:
      - application/json
      description: Updates group name.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Update group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpsertGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update group (admin)
      tags:
      - Admin Groups
  /admin/users:
    get:
      description: Returns paginated list of users.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List users (admin)
      tags:
      - Admin Users
  /admin/users/{id}:
    delete:
      description: Deletes user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Delete user (admin)
      tags:
      - Admin Users
    get:
      description: Returns user details by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get user by ID (admin)
      tags:
      - Admin Users
    patch:
      consumes:
      - application/json
      description: Partially updates user fields.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Admin update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update user (admin)
      tags:
      - Admin Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"